Feat: Add a small c program for adding/merging+more of Paseo steps
Instead of raw sql code or a python+dbt method,
this gives a small c program which transforms the paseo
output db and is able to merge new exports.

In addition, there is a TSV output which can be generated
through arguments to fetch some basic data.
This could be used to generate some simple plots.